+++ provider/pkg/gen/replacement.go
@@ -10,6 +10,13 @@ var forceNewMap = map[string]map[string]codegen.StringSet{
 	"Authorization": {
 		"RoleAssignment": codegen.NewStringSet("principalId", "scope"),
 	},
+	"Cdn": {
+		"Profile": codegen.NewStringSet(
+			"location",
+			// sku
+			"name",
+		),
+	},
 	"ContainerService": {
 		"ManagedCluster": codegen.NewStringSet(
 			// agentPoolProfiles